add swagger and dynamo connection, fix child class

+++ LambdaSample/LambdaSample.Middleware/Models/Child.cs
@@ -14,9 +14,9 @@
         public string? Type { get; set; }
         public string? Exchange_rate { get; set; }
         public string? Min_amount { get; set; }
-        public int Max_amount { get; set; }
+        public decimal Max_amount { get; set; }
         public string? Stock { get; set; }
-        public int Remaining_amount { get; set; }
+        public decimal Remaining_amount { get; set; }
         public DateTime Date_end { get; set; }
         public int Pips { get; set; }
         public decimal Discount { get; set; }
